Unified the scheduling page preview/public feature to one component
